requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==3.1.7
django-node==4.0.2
django-rest-knox==4.1.0
django-webpack-loader==0.7.0
django-webpack5-loader==0.8.1
djangorestframework==3.12.2
idna==2.10
lxml==4.6.3
optional-django==0.1.0
pkg-resources==0.0.0
pycparser==2.20
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4